api/requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.4.3; python_version >= '3.8'
aiohttp==3.11.2; python_version >= '3.9'
aiosignal==1.3.1; python_version >= '3.7'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.6.2.post1; python_version >= '3.9'
attrs==24.2.0; python_version >= '3.7'
blinker==1.9.0; python_version >= '3.9'
cachetools==5.5.0; python_version >= '3.7'
certifi==2024.8.30; python_version >= '3.6'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
dataclasses-json==0.6.7; python_version >= '3.7' and python_version < '4.0'
distro==1.9.0; python_version >= '3.6'
faiss-cpu==1.9.0; python_version >= '3.9'
fastapi==0.115.5; python_version >= '3.8'
frozenlist==1.5.0; python_version >= '3.8'
gitdb==4.0.11; python_version >= '3.7'
gitpython==3.1.43; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.27.2; python_version >= '3.8'
httpx-sse==0.4.0; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
jinja2==3.1.4; python_version >= '3.7'
jiter==0.7.1; python_version >= '3.8'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2024.10.1; python_version >= '3.9'
langchain==0.3.7; python_version >= '3.9' and python_version < '4.0'
langchain-community==0.3.7; python_version >= '3.9' and python_version < '4.0'
langchain-core==0.3.19; python_version >= '3.9' and python_version < '4.0'
langchain-openai==0.2.8; python_version >= '3.9' and python_version < '4.0'
langchain-text-splitters==0.3.2; python_version >= '3.9' and python_version < '4.0'
langsmith==0.1.143; python_full_version >= '3.8.1' and python_version < '4.0'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
marshmallow==3.23.1; python_version >= '3.9'
mdurl==0.1.2; python_version >= '3.7'
multidict==6.1.0; python_version >= '3.8'
mypy-extensions==1.0.0; python_version >= '3.5'
narwhals==1.13.5; python_version >= '3.8'
numpy==1.26.4; python_version >= '3.9'
openai==1.54.4; python_version >= '3.8'
orjson==3.10.11; python_version >= '3.8'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
pillow==11.0.0; python_version >= '3.9'
propcache==0.2.0; python_version >= '3.8'
protobuf==5.28.3; python_version >= '3.8'
pyarrow==18.0.0; python_version >= '3.9'
pydantic==2.9.2; python_version >= '3.8'
pydantic-core==2.23.4; python_version >= '3.8'
pydantic-settings==2.6.1; python_version >= '3.8'
pydeck==0.9.1; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
pypdf==5.1.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.1; python_version >= '3.8'
python-multipart==0.0.17; python_version >= '3.8'
pytz==2024.2
pyyaml==6.0.2; python_version >= '3.8'
referencing==0.35.1; python_version >= '3.8'
regex==2024.11.6; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rich==13.9.4; python_full_version >= '3.8.0'
rpds-py==0.21.0; python_version >= '3.9'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
smmap==5.0.1; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.35; python_version >= '3.7'
starlette==0.41.2; python_version >= '3.8'
tenacity==9.0.0; python_version >= '3.8'
tiktoken==0.8.0; python_version >= '3.9'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
tornado==6.4.1; python_version >= '3.8'
tqdm==4.67.0; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
typing-inspect==0.9.0
tzdata==2024.2; python_version >= '2'
urllib3==2.2.3; python_version >= '3.8'
uvicorn==0.32.0; python_version >= '3.8'
yarl==1.17.1; python_version >= '3.9'
